Change how archiving channels works
