+++ src/main/java/net/hypixel/nerdbot/api/database/model/user/stats/LastActivity.java
@@ -16,6 +16,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.function.ToLongFunction;
+import java.util.stream.Stream;
 
 @Getter
 @Setter
@@ -100,18 +101,16 @@ public class LastActivity {
     public List<ChannelActivityEntry> getChannelActivityHistory(int days) {
         LocalDate endDate = LocalDate.now();
         LocalDate startDate = endDate.minusDays(days - 1);
-
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("MM-yyyy");
 
         List<ChannelActivityEntry> entries = new ArrayList<>();
         for (ChannelActivityEntry entry : channelActivityHistory) {
             Map<String, Integer> monthlyMessageCountMap = entry.getMonthlyMessageCount();
-            for (LocalDate date = startDate; date.isBefore(endDate.plusDays(1)); date = date.plusDays(1)) {
-                String key = date.format(formatter);
-                if (monthlyMessageCountMap.containsKey(key)) {
-                    entries.add(new ChannelActivityEntry(entry.getChannelId(), entry.getLastKnownDisplayName(), monthlyMessageCountMap.get(key), entry.getLastMessageTimestamp(), monthlyMessageCountMap));
-                }
-            }
+            Stream.iterate(startDate, date -> date.isBefore(endDate.plusDays(1)), date -> date.plusDays(1))
+                .map(date -> date.format(formatter))
+                .filter(monthlyMessageCountMap::containsKey)
+                .map(key -> new ChannelActivityEntry(entry.getChannelId(), entry.getLastKnownDisplayName(), monthlyMessageCountMap.get(key), entry.getLastMessageTimestamp(), monthlyMessageCountMap))
+                .forEach(entries::add);
         }
 
         return entries;
